fix(SecuredEventInterceptor): Relocate by default for non-ajax events
For non-ajax events, `relocate` instead of `overrideEvent` to allow for
preProcess events and the like to fire completely when loading the
login page (or whatever page is the target of the redirect). For
ajax events, `overrideEvent` is still used.  This behavior can be
adjusted in the module settings.

BREAKING CHANGE: Non-ajax events now relocate instead of override the event.