Use local handler actions if they exists
So my proposal is that cbguard will check to see if in the handler itself, the relevant failure action that exists will be applied instead of the global overrides in the coldbox.cfc module settings. Giving us the ability to relocate on an individual handler basis. If it doesn't exists, the global overrides will fire.
This works with a simple app from https://github.com/elpete/quick-with-auth
I believe these changes will work but as i just started to look into cbguard, i don't have a complex app to test with.