feat(Guard): Add a guard service that can authorize anywhere during the request
The `Guard@cbguard` service can be used anywhere during the request (but
usually in a handler) to authorize a user against a `permission` and an
optional struct of `additionalArgs`.  This does not change any existing
aspects of cbguard.  In the future, a DSL may be introduced that allows
passing `additionalArgs` via the `secured` annotation.